Add files via upload
This code is all about printing the sum of each subarray.

Subarray: It is the continuous portion of the main array. That means subarray can be sliced from between but the sliced part should be of continuous indices else it will not be a subarray. The formula for no. of subarrays is (n*(n+1))/2.