Add files via upload
This code is all about the below question
Isyana is given the number of visitors at her local theme park on N consecutive days. The number of visitors on the i-th day is V.. 
A day is record breaking if it satisfies both of the following conditions:
The number of visitors on the day is strictly larger than the number of visitors on each of the previous days. Either it is the last day, or the number of visitors on the day is strictly larger than the number of visitors on the following day. Note that the very first day could be a record breaking day!
Please help Isyana find out the number of record breaking days.

Approach: using the max operator find the maximum element of every ith iteration and as we get the updated value of variable mx we will increase the counter which will tell the no. of days when the park received the record breaking days.