Add files via upload
This code is all about finding the smallest missing positive integer from an array.

Approach: Take another array of bool type and keep on checking the positive integers in the array and mark them as 1 or True as per your convenience. After the checking of all the elements re-iterate the bool array and check for the first 0 and that will be your answer.