Add files via upload
This code is all about the searching of element in the matrix but the catch here was the order of the elements of rows and columns should be in ascending order. One approach which can be the brute force approach in which we can use linear search and traverse every element. The time complexity for this approach would be O(nm). But here we didn't use the hint given in the question which was, the elements are in ascending order. 

In the optimised approach we can just pick two elements either from the top right corner or from the bottom left corner. In this code I have used top right corner, if we use the bottom left corner then the difference would come this much that if target is small then the element it is currently upon then we would do r-- else c++. Using this approach the time complexity is now reduced.