Sieve of Eratosthenes
WAP to implement the concept of Sieve of Eratosthenes through finding out the prime numbers in the entered range.

Sieve of Eratosthenes is a concept in which we mark the elements which are multiple of the prime numbers. Iterating through which at the end we are left with the unmarked elements which are actually the prime numbers.